Revert "moved AbstractWorkflowBasedLaunchConfigurationDelegate and AbstractConfigBuilderBasedLaunchConfigurationDelegate to de.uka.ipd.sdq.workflow.launchconfig.core"
This reverts commit d3757688ed7dfb61634df05146f5811db7eb8b2c.
